Documents/Dedolytics/crm/requirements.txt
google-generativeai==0.4.1
playwright==1.42.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
schedule==1.2.1
python-dotenv==1.0.1
pydantic==2.6.4
requests>=2.31.0
dnspython>=2.4.0
